Add ResponsiveCenter widget

+++ lib/core/presentation/widgets/responsive_widgets/responsive_layouts.dart
@@ -12,7 +12,7 @@ enum WindowSizeClass {
 
   const WindowSizeClass(this.breakpoint);
 
-  final int breakpoint;
+  final double breakpoint;
 
   static WindowSizeClass determineWindowClass(Size size) {
     // Use width rather than shortestSide when on Web/Desktop.
